# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Stock Management API",
        "description": "REST API untuk mengelola data inventaris, produk, supplier, warehouse, dan stok.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get All Products",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create Product",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "supplier_id",
                                    "warehouse_id",
                                    "name",
                                    "sku",
                                    "price"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop ASUS"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "SKU-001"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Gaming Laptop"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Product Detail",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update Product",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop ASUS ROG"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 17000000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Product"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    }
                }
            }
        },
        "/api/stock-in": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock In",
                "operationId": "531a3d78e6207ce47d91ba98b00374fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Barang masuk"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock In success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stock-out": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock Out",
                "operationId": "1fd2dae0e5d911259355e4791e8497cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Barang keluar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock Out success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stock-adjustment": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock Adjustment",
                "operationId": "43d64687699e68040b120735173706fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "stock"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Stock opname"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock adjusted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stock-history": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock History",
                "operationId": "9ec256033470c0f6f450e3c395646036",
                "responses": {
                    "200": {
                        "description": "List Stock History"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}